+++ test/CMakeLists.txt
@@ -10,6 +10,7 @@ add_unittest(binomial_coefficient)
 add_unittest(diophantine)
 add_unittest(lucas_sequence)
 add_unittest(matrix)
+add_unittest(modulo)
 add_unittest(pell_equation)
 add_unittest(prime_decomposition)
 add_unittest(prime_generation)
